Update README.md
🌐 Learning Experience
This project marks a significant milestone in my journey as a data analyst. It taught me:

How to design impactful dashboards tailored for business needs.
How to extract meaningful insights from data to drive decision-making.
The importance of storytelling in presenting data effectively.
🎉 Reflections
Completing this project with Power BI not only strengthened my technical skills but also deepened my understanding of the hospitality industry. It’s a testament to the power of data analytics in transforming businesses.